@@ -3,10 +3,6 @@
  * - newlines in text
  * - space in tts
  */
-const {textTts} = require('./text-tts');
+import {textTts} from './text-tts';
 
-const br = (n = 1) => textTts('\n'.repeat(n), ' ');
-
-module.exports = {
-  br,
-};
+export const br = (n = 1) => textTts('\n'.repeat(n), ' ');
