@@ -2,16 +2,12 @@
  * Puts first argument to `text` and second argument to `tts`.
  */
 
-const {text} = require('./text');
-const {tts} = require('./tts');
+import {text} from './text';
+import {tts} from './tts';
 
-const textTts = (textValue, ttsValue) => {
+export const textTts = (textValue, ttsValue) => {
   return {
     ...text(textValue),
     ...tts(ttsValue),
   };
-};
-
-module.exports = {
-  textTts,
 };
