@@ -2,22 +2,15 @@
  * Working with voice part.
  */
 
-const {stringify} = require('./utils');
-const {select} = require('./select');
+import {stringify} from './utils.js';
+import {select} from './select';
 
-const tts = value => {
+export const tts = (value) => {
   value = Array.isArray(value) ? select(value) : value;
   return {
-    tts: stringify(value)
+    tts: stringify(value),
   };
 };
-const audio = name => tts(`<speaker audio="alice-${name}.opus">`);
-const effect = name => tts(`<speaker effect="${name}">`);
-const pause = (ms = 500) => tts(`sil <[${ms}]>`);
-
-module.exports = {
-  tts,
-  audio,
-  effect,
-  pause,
-};
+export const audio = (name) => tts(`<speaker audio="alice-${name}.opus">`);
+export const effect = (name) => tts(`<speaker effect="${name}">`);
+export const pause = (ms = 500) => tts(`sil <[${ms}]>`);
