@@ -4,8 +4,8 @@
  * See: https://github.com/vitalets/alice-renderer/issues/1
  */
 
-const {isFunction, isObject} = require('./utils');
-const {getOrCreateSession, setUserId} = require('./sessions');
+import {isFunction, isObject} from './utils.js';
+import {getOrCreateSession, setUserId} from './sessions';
 
 /**
  * Userifies function or all methods of object.
@@ -14,7 +14,7 @@ const {getOrCreateSession, setUserId} = require('./sessions');
  * @param {function|object} target
  * @returns {function|object}
  */
-const userify = (userId, target) => {
+export const userify = (userId, target) => {
   void getOrCreateSession(userId);
   return isFunction(target)
     ? userifyFn(userId, target)
@@ -51,13 +51,7 @@ const userifyFn = (userId, fn) => {
 const userifyObj = (userId, obj) => {
   return new Proxy(obj, {
     get: (obj, prop) => {
-      return isFunction(obj[prop])
-        ? userifyFn(userId, obj[prop])
-        : obj[prop];
-    }
+      return isFunction(obj[prop]) ? userifyFn(userId, obj[prop]) : obj[prop];
+    },
   });
-};
-
-module.exports = {
-  userify,
 };
