@@ -2,7 +2,7 @@
  * Configuration method.
  */
 
-const config = {
+export const config = {
   /**
    * Disables randomization, all replies does not vary (useful for tests)
    */
@@ -15,17 +15,12 @@ const config = {
  * @param {Object} options
  * @param {Boolean} options.disableRandom
  */
-const configure = options => {
-  Object.keys(options || {}).forEach(key => {
+export const configure = (options) => {
+  Object.keys(options || {}).forEach((key) => {
     if (config.hasOwnProperty(key)) {
       config[key] = options[key];
     } else {
       throw new Error(`Unknown option: ${key}`);
     }
   });
-};
-
-module.exports = {
-  configure,
-  config,
 };
