@@ -10,24 +10,24 @@
  * @param {string} [lastSeparator]
  * @returns {string}
  */
-const enumerate = (arr, { separator = ', ', lastSeparator = ' или ' } = {}) => {
+export const enumerate = (
+  arr,
+  {separator = ', ', lastSeparator = ' или '} = {}
+) => {
   if (!Array.isArray(arr)) {
     throw new Error(`You should pass array in enumerate()`);
   }
   arr = arr.filter(Boolean);
   switch (arr.length) {
-    case 0: return '';
-    case 1: return arr[0];
+    case 0:
+      return '';
+    case 1:
+      return arr[0];
     default: {
       const [prev, last] = arr.slice(-2);
-      return [
-        ...arr.slice(0, -2),
-        `${prev}${lastSeparator}${last}`
-      ].join(separator);
+      return [...arr.slice(0, -2), `${prev}${lastSeparator}${last}`].join(
+        separator
+      );
     }
   }
-};
-
-module.exports = {
-  enumerate,
 };
