@@ -1,4 +1,6 @@
-const { enumerate } = require('../../');
+require = require('esm')(module);
+
+const {enumerate} = require('../../src');
 
 describe('enumerate', () => {
 
@@ -9,17 +11,17 @@ describe('enumerate', () => {
   });
 
   it('0 elements', () => assert.equal(enumerate([]), ''));
-  it('1 element', () => assert.equal(enumerate([ 'foo' ]), 'foo'));
-  it('2 elements', () => assert.equal(enumerate([ 'foo', 'bar' ]), 'foo или bar'));
-  it('3 elements', () => assert.equal(enumerate([ 'foo', 'bar', 'baz' ]), 'foo, bar или baz'));
+  it('1 element', () => assert.equal(enumerate(['foo']), 'foo'));
+  it('2 elements', () => assert.equal(enumerate(['foo', 'bar']), 'foo или bar'));
+  it('3 elements', () => assert.equal(enumerate(['foo', 'bar', 'baz']), 'foo, bar или baz'));
 
   it('filter falsy elements', () => {
-    const res = enumerate([ undefined, 'foo', false, 'bar', null, 'baz' ]);
+    const res = enumerate([undefined, 'foo', false, 'bar', null, 'baz']);
     assert.equal(res, 'foo, bar или baz');
   });
 
   it('custom separators', () => {
-    const res = enumerate([ 'foo', 'bar', 'baz' ], { separator: ' или ', lastSeparator: ', а также ' });
+    const res = enumerate(['foo', 'bar', 'baz'], {separator: ' или ', lastSeparator: ', а также '});
     assert.equal(res, 'foo или bar, а также baz');
   });
 });
