@@ -1,3 +1,5 @@
+require = require('esm')(module);
+
 const {reply, image} = require('../../src');
 
 describe('image: appendDescription', () => {
@@ -6,7 +8,7 @@ describe('image: appendDescription', () => {
     const text = 'а'.repeat(128);
     const res = reply`
       ${text}
-      ${image('1234567/xxx', { appendDescription: 'foo' })}
+      ${image('1234567/xxx', {appendDescription: 'foo'})}
     `;
     assert.equal(res.text, text);
     assert.deepEqual(res.card, {
@@ -21,7 +23,7 @@ describe('image: appendDescription', () => {
     const text = 'а'.repeat(200);
     const res = reply`
       ${text}
-      ${image('1234567/xxx', { appendDescription: 'foo' })}
+      ${image('1234567/xxx', {appendDescription: 'foo'})}
     `;
     assert.deepEqual(res.card, {
       type: 'BigImage',
@@ -31,13 +33,13 @@ describe('image: appendDescription', () => {
     });
   });
 
-  it('text > 256: write to title and description with appendDescription',  () => {
+  it('text > 256: write to title and description with appendDescription', () => {
     const titleChunk = 'Ку-ку. Привет Привет Привет!';
     const descriptionChunk = 'А'.repeat(256 - titleChunk.length);
     const res = reply`
       ${titleChunk}
       ${descriptionChunk}
-      ${image('1234567/xxx', { appendDescription: 'foo' })}
+      ${image('1234567/xxx', {appendDescription: 'foo'})}
     `;
     assert.deepEqual(res.card, {
       type: 'BigImage',
@@ -51,7 +53,7 @@ describe('image: appendDescription', () => {
     const appendDescription = 'а'.repeat(300);
     const res = reply`
       text
-      ${image('1234567/xxx', { appendDescription })}
+      ${image('1234567/xxx', {appendDescription})}
     `;
     assert.equal(res.text, 'text');
     assert.deepEqual(res.card, {
